Make the pose_samples drive the PIDController
A new cmd_out is sent only when there is a new pose_sample.
The timestamp of the cmd_out is the same of the pose_sample.
It happens cause the cmd_in, being the set-point of the controller,
doesn't need to be received all time. In the other handy, the pose_sample
being the feedback of the system need to be compared constantly in order
to generate a new cmd_out.