Create POSE_TIMEOUT exception for WorldToAligned
Create also ORIENTATION_TIMEOUT exception for AlignedToBody.

Define WAIT_FOR_POSE/ORIENTATION_SAMPLE as runtime states,
once the perspectives tasks wait for the first data in that state.
Once it's an OldData for pose, the tasks keep in CONTROLLING
(using the last pose) until it reach the TIMEOUT exception.

As both tasks do a frame transformation between gatheredInputCommand
to cmd_out, there is only a cmd_out for a NEW_COMMAND in gatherInputCommand