Refactor Base
Create OneLoop that embedded gatherInputCommand and calcOutPut.
OneLoop is called in both Base::runtimeHook and Base::errorHook.
The State transition is done in both Hook, according the state returned by
OneLoop.
In case the derived class should switch to its own error/runtime state,
it should be done in the re-implementation of calcOutput and return false.

Create enum CommandStatus and make gatherInputCommand assign both
the command and its status in LinearAngular6DCommandStatus. Make them be
the argument for calcOutput.