Code-refactoring in BasePIDController and PIDController
Create structs for putting together linear and angular PIDs, currentState
and currentCov.
Embedded timestamp in LinearAngular6DPIDState, once it's present as
an output_port.

Define specific timeout_pose in BasePIDController and create
calcPIDStateCommand. It's called in calcOutput and requires
reference (merged_command), the current_state (pose_sample) and the PIDs
(they are updated). It provides the command_output and the pid_state and
their timestamp are set from the current_state.